Apply review suggestions + improve unwrapping
Add support for unwrapping keys with .@each and .[] along with multiple
nesting
