+++ lib/rules/no-duplicate-dependent-keys.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const utils = require('../utils/utils');
+const emberUtils = require('../utils/ember');
 
 const MESSAGE = 'Dependent keys should not be repeated';
 //------------------------------------------------------------------------------
@@ -21,7 +21,7 @@ module.exports = {
   create(context) {
     return {
       CallExpression(node) {
-        if (utils.hasDuplicateDependentKeys(node)) {
+        if (emberUtils.hasDuplicateDependentKeys(node)) {
           context.report(node, MESSAGE);
         }
       }
