+++ lib/utils/ember.js
@@ -42,6 +42,9 @@ module.exports = {
 
   isControllerProperty,
   isControllerDefaultProp,
+  parseDependentKeys,
+  unwrapBraceExpressions,
+  hasDuplicateDependentKeys,
 };
 
 // Private
@@ -350,4 +353,69 @@ function isRelation(property) {
   });
 
   return result;
+}
+
+/**
+ * Checks whether a computed property has duplicate dependent keys.
+ *
+ * @param  {CallExpression} callExp Given call expression
+ * @return {Boolean}        Flag whether dependent keys present.
+ */
+function hasDuplicateDependentKeys(callExp) {
+  if (!isComputedProp(callExp)) return false;
+
+  const dependentKeys = parseDependentKeys(callExp);
+  const uniqueKeys = dependentKeys
+    .filter((val, index, self) => self.indexOf(val) === index);
+
+  return uniqueKeys.length !== dependentKeys.length;
+}
+
+/**
+ * Parses dependent keys from call expression and returns them in an array.
+ *
+ * It also unwraps the expressions, so that `model.{foo,bar}` becomes `model.foo, model.bar`.
+ *
+ * @param  {CallExpression} callExp CallExpression to examine
+ * @return {Array}          Array of unwrapped dependent keys
+ */
+function parseDependentKeys(callExp) {
+  // Check whether we have a MemberExpression, eg. computed(...).volatile()
+  const isMemberExpCallExp = !callExp.arguments.length &&
+    utils.isMemberExpression(callExp.callee) &&
+    utils.isCallExpression(callExp.callee.object);
+
+  const args = isMemberExpCallExp ? callExp.callee.object.arguments : callExp.arguments;
+
+  const dependentKeys = args
+    .filter(arg => utils.isLiteral(arg))
+    .map(literal => literal.value);
+
+  return unwrapBraceExpressions(dependentKeys);
+}
+
+/**
+ * Unwraps brace expressions.
+ *
+ * @param  {Array} dependentKeys array of strings containing unprocessed dependent keys.
+ * @return {Array} Array of unwrapped dependent keys
+ */
+function unwrapBraceExpressions(dependentKeys) {
+  const braceExpressionRegexp = /{.+}/g;
+
+  const unwrappedExpressions = dependentKeys.map((key) => {
+    if (!braceExpressionRegexp.test(key)) return key;
+
+    const braceExpansionPart = key.match(braceExpressionRegexp)[0];
+    const prefix = key.replace(braceExpansionPart, '');
+    const properties = braceExpansionPart
+      .replace('{', '')
+      .replace('}', '')
+      .split(',');
+
+    return properties.map(property => `${prefix}${property}`);
+  });
+
+  return unwrappedExpressions
+    .reduce((acc, cur) => acc.concat(cur), []);
 }
