+++ tests/lib/utils/utils-test.js
@@ -186,86 +186,3 @@ describe('parseArgs', () => {
     expect(parsedArgs).toEqual(['asd', 'qwe', 'zxc']);
   });
 });
-
-describe('parseDependentKeys', () => {
-  it('should parse dependent keys from callexpression', () => {
-    const node = parse("computed('model.{foo,bar}', 'model.bar')");
-    expect(utils.parseDependentKeys(node)).toEqual([
-      'model.foo', 'model.bar', 'model.bar',
-    ]);
-  });
-
-  it('should work when no dependent keys present', () => {
-    const node = parse('computed(function() {})');
-    expect(utils.parseDependentKeys(node)).toEqual([]);
-  });
-
-  it('should handle dependent keys and function arguments', () => {
-    const node = parse("computed('model.{foo,bar}', 'model.bar', function() {})");
-    expect(utils.parseDependentKeys(node)).toEqual([
-      'model.foo', 'model.bar', 'model.bar',
-    ]);
-  });
-
-  it('should handle dependent keys and function arguments in MemberExpression', () => {
-    const node = parse(`
-      computed('model.{foo,bar}', 'model.bar', function() {
-      }).volatile();
-    `);
-    expect(utils.parseDependentKeys(node)).toEqual([
-      'model.foo', 'model.bar', 'model.bar',
-    ]);
-  });
-});
-
-describe('unwrapBraceExpressions', () => {
-  it('should unwrap simple dependent keys', () => {
-    expect(utils.unwrapBraceExpressions([
-      'model.foo', 'model.bar'
-    ])).toEqual(['model.foo', 'model.bar']);
-  });
-
-  it('should unwrap dependent keys with braces', () => {
-    expect(utils.unwrapBraceExpressions([
-      'model.{foo,bar}', 'model.bar'
-    ])).toEqual(['model.foo', 'model.bar', 'model.bar']);
-  });
-
-  it('should unwrap more complex dependent keys', () => {
-    expect(utils.unwrapBraceExpressions([
-      'model.{foo,bar}', 'model.bar', 'data.{foo,baz,qux}'
-    ])).toEqual([
-      'model.foo', 'model.bar', 'model.bar', 'data.foo', 'data.baz', 'data.qux',
-    ]);
-  });
-});
-
-describe('hasDuplicateDependentKeys', () => {
-  it('reports duplicate dependent keys in computed calls', () => {
-    let node = parse("computed('model.{foo,bar}', 'model.bar')");
-    expect(utils.hasDuplicateDependentKeys(node)).toBeTruthy();
-    node = parse("Ember.computed('model.{foo,bar}', 'model.bar')");
-    expect(utils.hasDuplicateDependentKeys(node)).toBeTruthy();
-  });
-
-  it('ignores not repeated dependentKeys', () => {
-    let node = parse("computed('model.{foo,bar}', 'model.qux')");
-    expect(utils.hasDuplicateDependentKeys(node)).not.toBeTruthy();
-    node = parse("Ember.computed('model.{foo,bar}', 'model.qux')");
-    expect(utils.hasDuplicateDependentKeys(node)).not.toBeTruthy();
-    node = parse("computed('model.{foo,bar}', 'model.qux').volatile()");
-    expect(utils.hasDuplicateDependentKeys(node)).not.toBeTruthy();
-  });
-
-  it('ignores non-computed calls', () => {
-    const node = parse("foo('model.{foo,bar}', 'model.bar')");
-    expect(utils.hasDuplicateDependentKeys(node)).not.toBeTruthy();
-  });
-
-  it('reports duplicate dependent keys in computed calls with MemberExp', () => {
-    let node = parse("Ember.computed('model.{foo,bar}', 'model.bar').volatile()");
-    expect(utils.hasDuplicateDependentKeys(node)).toBeTruthy();
-    node = parse("computed('model.{foo,bar}', 'model.bar').volatile()");
-    expect(utils.hasDuplicateDependentKeys(node)).toBeTruthy();
-  });
-});
